fix(Safari): Rewrite using a different techniqe
Checking with querySelector(':focus-within') only works if the element has focus.
Unfortunately, according to the spec `onBlur` event occurs after the blur already happened
so the selector always returns `null` in Safari.

This commit rewrites the logic to not rely on the CSS selector anymore. This also makes it
work in browsers that doesn't support `:focus-within` yet without the need of polyfill.

Fixes #4